requirements.txt
Flask==2.3.3
requests==2.31.0
pandas==2.1.0
plotly==5.17.0
python-dotenv==1.0.0
apscheduler==3.10.4 